🚚 Move InputSource into its own module
Continuing the effort to make `input::mod` smaller, and only re-export.
