+++ src/input/source.rs
@@ -1,0 +1,8 @@
+use crate::input::InputToken;
+
+pub trait InputSource {
+	type Token: InputToken;
+	fn source_name(&self) -> String;
+	fn next_token(&mut self) -> Option<Self::Token>;
+	fn peek(&mut self) -> Option<&Self::Token>;
+}
